pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "product-management"
version = "0.1.0"
description = "CLI-first system for managing product ideas through autonomous LLM execution"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-frontmatter>=1.0",
    "anthropic>=0.39",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
pm = "pm.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pm"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
